refactor: generalize web gui right panel
