+++ web-gui/app/src/app/App.tsx
@@ -6,7 +6,7 @@ import { EmptyState } from "../components/ui/EmptyState";
 import { SegmentedControl, SegmentedControlButton } from "../components/ui/SegmentedControl";
 import { StatusBadge } from "../components/ui/StatusChip";
 import { DashboardPage } from "../features/dashboard/DashboardPage";
-import { InspectorPanel } from "../features/inspector/InspectorPanel";
+import { RightSidePanel } from "../features/right-panel/RightSidePanel";
 import { SearchPage } from "../features/search/SearchPage";
 import { SettingsPage } from "../features/settings/SettingsPage";
 import { deriveAgentDisplayStatus } from "../runtime/agent-status";
@@ -28,16 +28,16 @@ export function App() {
   const route = useRuntimeStore((state) => state.route);
   const selectedAgentId = useRuntimeStore((state) => state.selectedAgentId);
   const displayLevel = useRuntimeStore((state) => state.displayLevel);
-  const inspectorOpen = useRuntimeStore((state) => state.inspectorOpen);
-  const inspectorSelection = useRuntimeStore((state) => state.inspectorSelection);
+  const rightPanelOpen = useRuntimeStore((state) => state.rightPanelOpen);
+  const rightPanelView = useRuntimeStore((state) => state.rightPanelView);
   const navCollapsed = useRuntimeStore((state) => state.navCollapsed);
   const setRoute = useRuntimeStore((state) => state.setRoute);
   const openAgent = useRuntimeStore((state) => state.openAgent);
   const setDisplayLevel = useRuntimeStore((state) => state.setDisplayLevel);
-  const setInspectorOpen = useRuntimeStore((state) => state.setInspectorOpen);
+  const setRightPanelOpen = useRuntimeStore((state) => state.setRightPanelOpen);
   const inspectActivity = useRuntimeStore((state) => state.inspectActivity);
-  const clearInspectorSelection = useRuntimeStore((state) => state.clearInspectorSelection);
-  const toggleInspector = useRuntimeStore((state) => state.toggleInspector);
+  const showAgentOverview = useRuntimeStore((state) => state.showAgentOverview);
+  const toggleRightPanel = useRuntimeStore((state) => state.toggleRightPanel);
   const toggleNavCollapsed = useRuntimeStore((state) => state.toggleNavCollapsed);
   const setRuntimeConnection = useRuntimeStore((state) => state.setRuntimeConnection);
   const selectedAgent = useRuntimeStore(selectSelectedAgent);
@@ -174,7 +174,7 @@ export function App() {
   return (
     <div
       className="app-shell"
-      data-panel={inspectorOpen ? "open" : "closed"}
+      data-panel={rightPanelOpen ? "open" : "closed"}
       data-nav-collapsed={navCollapsed}
     >
       <aside className="sidebar" aria-label="Holon navigation">
@@ -288,9 +288,9 @@ export function App() {
                 type="button"
                 size="icon"
                 variant="ghost"
-                aria-label="Toggle object inspector"
-                title="Toggle object inspector"
-                onClick={toggleInspector}
+                aria-label="Toggle context panel"
+                title="Toggle context panel"
+                onClick={toggleRightPanel}
               >
                 ▭
               </Button>
@@ -329,13 +329,12 @@ export function App() {
             onLoadOlderEvents={() => loadOlderAgentEvents(activeAgent.id, displayLevel)}
             onSendPrompt={(text) => sendOperatorPrompt(activeAgent.id, text, displayLevel)}
             onOpenInspector={() => {
-              clearInspectorSelection();
-              setInspectorOpen(true);
+              showAgentOverview(activeAgent.id);
             }}
             onInspectActivity={(activity) => inspectActivity(activeAgent.id, activity)}
             selectedActivityId={
-              inspectorSelection?.kind === "activity" && inspectorSelection.agentId === activeAgent.id
-                ? inspectorSelection.activity.id
+              rightPanelView?.kind === "activity_inspector" && rightPanelView.agentId === activeAgent.id
+                ? rightPanelView.activity.id
                 : undefined
             }
           />
@@ -374,12 +373,12 @@ export function App() {
       </main>
 
       {selectedAgent ? (
-        <InspectorPanel
+        <RightSidePanel
           agent={selectedAgent}
-          selection={inspectorSelection?.agentId === selectedAgent.id ? inspectorSelection : undefined}
-          open={inspectorOpen}
-          onClearSelection={clearInspectorSelection}
-          onClose={() => setInspectorOpen(false)}
+          view={rightPanelView?.agentId === selectedAgent.id ? rightPanelView : undefined}
+          open={rightPanelOpen}
+          onShowAgentOverview={showAgentOverview}
+          onClose={() => setRightPanelOpen(false)}
         />
       ) : null}
     </div>
